Resolving AuthenticationProvider WARN
We don't need at the moment to define an authenticationProvider because
we have only one UserDetailsService implementation
(org.isf.security.UserDetailsServiceImpl) and only one PasswordEncoder
bean (BCryptPasswordEncoder), so Spring Security uses the default
DaoAuthenticationProvider

In the future, if we add more UserDetailsService implementations for
multiple authentication data sources (e.g. Database, LDAP, Third-Party)
we must define each provider and instruct the AuthenticationManager in
which one to use for each case.